Dedicated directory for tests.
