+++ packages/outbox-core/test/outbox.spec.ts
@@ -12,10 +12,9 @@ import {
 import pino, { type Logger } from 'pino'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 import { z } from 'zod'
-import { InMemoryOutboxAccumulator, type OutboxAccumulator } from './accumulators'
-import type { OutboxEntry } from './objects'
-import { type OutboxDependencies, OutboxEventEmitter, OutboxProcessor } from './outbox'
-import type { OutboxStorage } from './storage'
+import { InMemoryOutboxAccumulator } from '../lib/accumulators'
+import { type OutboxDependencies, OutboxEventEmitter, OutboxProcessor } from '../lib/outbox'
+import { InMemoryOutboxStorage } from './InMemoryOutboxStorage'
 
 const TestEvents = {
   created: {
@@ -53,76 +52,6 @@ const createdEventPayload: CommonEventDefinitionPublisherSchemaType<typeof TestE
 }
 
 const TestLogger: Logger = pino()
-
-class InMemoryOutboxStorage<SupportedEvents extends CommonEventDefinition[]>
-  implements OutboxStorage<SupportedEvents>
-{
-  public entries: OutboxEntry<SupportedEvents[number]>[] = []
-
-  createEntry(
-    outboxEntry: OutboxEntry<SupportedEvents[number]>,
-  ): Promise<OutboxEntry<SupportedEvents[number]>> {
-    this.entries = [...this.entries, outboxEntry]
-
-    return Promise.resolve(outboxEntry)
-  }
-
-  getEntries(maxRetryCount: number): Promise<OutboxEntry<SupportedEvents[number]>[]> {
-    const entries = this.entries.filter((entry) => {
-      return entry.status !== 'SUCCESS' && entry.retryCount <= maxRetryCount
-    })
-
-    return Promise.resolve(entries)
-  }
-
-  update(
-    outboxEntry: OutboxEntry<SupportedEvents[number]>,
-  ): Promise<OutboxEntry<SupportedEvents[number]>> {
-    this.entries = this.entries.map((entry) => {
-      if (entry.id === outboxEntry.id) {
-        return outboxEntry
-      }
-      return entry
-    })
-
-    return Promise.resolve(outboxEntry)
-  }
-
-  public async flush(outboxAccumulator: OutboxAccumulator<SupportedEvents>): Promise<void> {
-    let successEntries = await outboxAccumulator.getEntries()
-    successEntries = successEntries.map((entry) => {
-      return {
-        ...entry,
-        status: 'SUCCESS',
-        updateAt: new Date(),
-      }
-    })
-    this.entries = this.entries.map((entry) => {
-      const foundEntry = successEntries.find((successEntry) => successEntry.id === entry.id)
-      if (foundEntry) {
-        return foundEntry
-      }
-      return entry
-    })
-
-    let failedEntries = await outboxAccumulator.getFailedEntries()
-    failedEntries = failedEntries.map((entry) => {
-      return {
-        ...entry,
-        status: 'FAILED',
-        updateAt: new Date(),
-        retryCount: entry.retryCount + 1,
-      }
-    })
-    this.entries = this.entries.map((entry) => {
-      const foundEntry = failedEntries.find((failedEntry) => failedEntry.id === entry.id)
-      if (foundEntry) {
-        return foundEntry
-      }
-      return entry
-    })
-  }
-}
 
 const MAX_RETRY_COUNT = 2
